+++ docs/changelog.rst
@@ -16,6 +16,8 @@ New Features
 - Use example c++ library in Abaqus UMAT template (:jira:`505`, :pull:`8`). By `Kyle Brindley`_.
 - Add c++ to fortran variable conversion and Abaqus variable return template (:jira:`521`, :pull:`15`, :pull:`16`). By
   `Kyle Brindley`_.
+- Add common abaqus tensor handling tools and a c++ converted umat interface (:jira:`522`, :pull:`17`). By `Kyle
+  Brindley`_.
 
 Bug fixes
 ~~~~~~~~~
@@ -32,6 +34,8 @@ Internal Changes
 - Update Jenkins CI configuration and store with version controlled repository (:jira:`442`, :pull:`5`). By `Kyle Brindley`_.
 - Demonstrate c++ ``vector_tools`` library for unit testing (:jira:`506`, :pull:`7`). By `Kyle Brindley`_.
 - Add integration tests for Abaqus UMAT interface (:jira:`504`, :pull:`10`). By `Kyle Brindley`_.
+- Move project Abaqus interface into project files. Treat UMAT Fortran/c++ subroutine as a UMAT selection and pass
+  through subroutine (:jira:`523`, :pull:`18`). By `Kyle Brindley`_.
 
 Enhancements
 ~~~~~~~~~~~~